refactor selectOneMenu to own class (use selenide)
