+++ primeui-tester/src/test/java/com/axonivy/ivy/supplements/primeui/tester/TestPrimeUi.java
@@ -16,10 +16,10 @@ import org.openqa.selenium.support.ui.ExpectedConditions;
 import com.axonivy.ivy.supplements.primeui.tester.PrimeUi.Accordion;
 import com.axonivy.ivy.supplements.primeui.tester.PrimeUi.Dialog;
 import com.axonivy.ivy.supplements.primeui.tester.PrimeUi.SelectBooleanCheckbox;
-import com.axonivy.ivy.supplements.primeui.tester.PrimeUi.SelectCheckboxMenu;
-import com.axonivy.ivy.supplements.primeui.tester.PrimeUi.SelectOneMenu;
 import com.axonivy.ivy.supplements.primeui.tester.PrimeUi.SelectOneRadio;
 import com.axonivy.ivy.supplements.primeui.tester.PrimeUi.Table;
+import com.axonivy.ivy.supplements.primeui.tester.widget.SelectCheckboxMenu;
+import com.axonivy.ivy.supplements.primeui.tester.widget.SelectOneMenu;
 import com.codeborne.selenide.Configuration;
 import com.codeborne.selenide.Selenide;
 import com.codeborne.selenide.WebDriverRunner;
@@ -39,7 +39,7 @@ public class TestPrimeUi
   public void setUp()
   {
     Configuration.browser = "firefox";
-    Configuration.headless = true;
+    //Configuration.headless = true;
     Configuration.reportsFolder = "target/senenide/reports";
     Selenide.open();
     driver = WebDriverRunner.getWebDriver();
